Revisione dashboard, aggiunto hash to room. Revisione test rspec & cucumber. Testati calendar in user, join the queue, e set date
