+++ app/assets/javascripts/cable.js
@@ -4,10 +4,3 @@
 //= require action_cable
 //= require_self
 //= require_tree ./channels
-
-(function() {
-  this.App || (this.App = {});
-
-  App.cable = ActionCable.createConsumer();
-
-}).call(this);
