+++ app/assets/stylesheets/mapbox.scss
@@ -10,7 +10,21 @@
 }
 
 .mapboxgl-ctrl-geocoder{
-  width:100%;
-  min-width: 260px;
+  width:105%;
+  min-width: none;
   height: 36px !important;
 }
+
+.card_body{
+  position: relative;
+}
+
+.card-link{
+  position: absolute;
+  bottom: 10px;
+}
+.card-img-top{
+  width: 100%;
+  height: 200px;
+}
+
