+++ app/assets/stylesheets/rooms.scss
@@ -13,6 +13,11 @@
   width:24px;
 }
 
+.mapboxgl-ctrl-geocoder{
+  width: 100%;
+  max-width: none;
+}
+
 .body-room{
   height: 70vh;
 }
