+++ app/controllers/application_controller.rb
@@ -1,11 +1,11 @@
 class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
-
-  #rescue_from ActiveRecord::RecordNotFound, with: :render_404
+  before_action :set_no_cache
+  rescue_from ActiveRecord::RecordNotFound, with: :render_404
   
   include SessionsHelper
   helper_method :current_user
-
+  
   def current_user
     @current_user ||= User.find(session[:user_id]) if session[:user_id]
   end
@@ -15,5 +15,11 @@ class ApplicationController < ActionController::Base
       format.html { render file: "#{Rails.root}/public/404.html", layout: 'layouts/application', status: 404 }
       format.all { render nothing: true, status: 404 }
     end
+  end
+
+  def set_no_cache
+    response.headers["Cache-Control"] = "no-cache, no-store, max-age=0, must-revalidate"
+    response.headers["Pragma"] = "no-cache"
+    response.headers["Expires"] = "Fri, 01 Jan 1990 00:00:00 GMT"
   end
 end
