+++ app/controllers/rooms_controller.rb
@@ -1,4 +1,5 @@
 class RoomsController < ApplicationController
+  before_action :set_room, only: [:edit, :update, :destroy, :destroy_avatar]
   before_action :logged_in_user, only: [:new]
   before_action :correct_user, only: [:edit, :destroy]
   
@@ -11,7 +12,7 @@ class RoomsController < ApplicationController
   # GET /rooms/1
   # GET /rooms/1.json
   def show
-    @room = Room.find(params[:id])
+    @room = Room.friendly.find(params[:id])
     @reservations = @room.reservations
   end
 
@@ -22,7 +23,6 @@ class RoomsController < ApplicationController
 
   # GET /rooms/1/edit
   def edit
-    @room = Room.find(params[:id])
   end
 
   # POST /rooms
@@ -47,7 +47,6 @@ class RoomsController < ApplicationController
   # PATCH/PUT /rooms/1
   # PATCH/PUT /rooms/1.json
   def update
-    @room = Room.find(params[:id])
     respond_to do |format|
       if @room.update(room_params)
         flash[:success] = 'Room updated!'
@@ -64,8 +63,6 @@ class RoomsController < ApplicationController
   # DELETE /rooms/1
   # DELETE /rooms/1.json
   def destroy
-    
-    @room = Room.find(params[:id]) if @room.nil?
     @room.destroy
     respond_to do |format|
       format.html { redirect_to rooms_url, notice: 'Room was successfully destroyed.' }
@@ -74,7 +71,6 @@ class RoomsController < ApplicationController
   end
 
   def destroy_avatar
-    @room = Room.find(params[:id])
     @room.avatar.destroy
     @room.save!
     redirect_to edit_room_path(@room)
@@ -83,7 +79,7 @@ class RoomsController < ApplicationController
   private
     # Use callbacks to share common setup or constraints between actions.
     def set_room
-      @room = Room.find(params[:id])
+      @room = Room.friendly.find(params[:id])
     end
 
     # Never trust parameters from the scary internet, only allow the white list through.
@@ -100,7 +96,6 @@ class RoomsController < ApplicationController
     end
     
      def correct_user
-      @room = Room.find(params[:id])
       redirect_to(root_url) unless (!current_user.nil? && current_user.powers.exists?(room_id: @room.id))
     end
 end
