+++ app/helpers/reservation_helper.rb
@@ -1,6 +1,6 @@
 module ReservationHelper
    def can_delete?
-    room = Room.find(params[:id])
+    room = Room.friendly.find(params[:id])
     DateTime.current < room.max_unjoin_time
   end
 end
