+++ app/views/layouts/application.html.haml
@@ -10,13 +10,13 @@
    
     <script src="https://cdn.jsdelivr.net/npm/flatpickr"></script>
     <script src='https://npmcdn.com/@turf/turf/turf.min.js'></script>
-    /<link href='https://api.tiles.mapbox.com/mapbox-gl-js/v0.45.0/mapbox-gl.css' rel='stylesheet' />
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/flatpickr/dist/flatpickr.min.css">
     %script{src: "https://api.mapbox.com/mapbox-gl-js/plugins/mapbox-gl-geocoder/v2.2.0/mapbox-gl-geocoder.min.js"}
     %link{href: "https://api.mapbox.com/mapbox-gl-js/plugins/mapbox-gl-geocoder/v2.2.0/mapbox-gl-geocoder.css", rel: "stylesheet", type: "text/css"}/
     = stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track': 'reload'
     = javascript_include_tag 'application', 'data-turbolinks-track': 'reload'
     = stylesheet_link_tag 'navbar'
+
   %body
     =render "layouts/navbar"
     .container#flash_div
@@ -24,4 +24,3 @@
         %div{class: "alert alert-#{message_type}"}= message
     = yield
     /= render 'layouts/footer'
-    = debug(params) if Rails.env.development?
