+++ app/views/rooms/_room.json.jbuilder
@@ -1,2 +1,2 @@
-json.extract! room, :id, :name, :description, :private, :address, :user_id, :max_participants, :latitude, :longitude, :time_from, :time_to, :avatar, :recurrence, :event_id
+json.extract! room, :hash_id, :name, :description, :private, :address, :user_id, :max_participants, :latitude, :longitude, :time_from, :time_to, :avatar, :recurrence, :event_id
 json.url room_url(room, format: :json)
