+++ db/migrate/20180702234121_add_hash_id_to_rooms.rb
@@ -1,0 +1,6 @@
+class AddHashIdToRooms < ActiveRecord::Migration[5.2]
+  def change
+    add_column :rooms, :hash_id, :string, index: true
+    Room.all.each{|m| m.set_hash_id; m.save}
+  end
+end
