+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2018_06_25_230738) do
+ActiveRecord::Schema.define(version: 2018_07_02_234121) do
 
   create_table "messages", force: :cascade do |t|
     t.integer "room_id"
@@ -65,6 +65,7 @@ ActiveRecord::Schema.define(version: 2018_06_25_230738) do
     t.string "avatar_content_type"
     t.integer "avatar_file_size"
     t.datetime "avatar_updated_at"
+    t.string "hash_id"
     t.index ["user_id"], name: "index_rooms_on_user_id"
   end
