--- features/step_definitions/user_step.rb
@@ -1,2 +1,0 @@
-
-
