+++ features/support/paths.rb
@@ -1,16 +1,6 @@
 Before do |scenario|
-  @port = Capybara.current_session.server.port
+  @port = 5555
+  Capybara.server_port = @port
   @test_url = "http://127.0.0.1:#{@port}/"
-  Capybara.default_max_wait_time = 7
-end
-
-After do |scenario|
-  puts "Procedo all'eliminazione dei residui dello scenario"
-  cal = Inline::Application.config.cal
-  cal.delete_event('primary', @room.event_id) unless !@room
-  
-  #if scenario.failed?
-  #  subject = "[Project X] #{scenario.exception.message}"
-  #  send_failure_email(subject)
-  #end
+  Capybara.default_max_wait_time = 30
 end
